test: move custom context concerns into separate file
