+++ test/test-feel.js
@@ -1,14 +1,23 @@
 import { expect } from 'chai';
 
-import { parser, trackVariables, VariableContext } from 'lezer-feel';
 import { testTree } from '@lezer/generator/dist/test';
 import { buildParser } from '@lezer/generator';
+
+import {
+  parser,
+  trackVariables
+} from 'lezer-feel';
 
 import fs from 'node:fs';
 import path from 'node:path';
 
 import { fileURLToPath } from 'node:url';
 import { ContextTracker } from '@lezer/lr';
+
+import {
+  EntriesContext,
+  toEntriesContextValue
+} from './custom-context.js';
 
 const caseDir = path.dirname(fileURLToPath(import.meta.url));
 
@@ -229,113 +238,4 @@ for (const file of fs.readdirSync(caseDir)) {
 
   });
 
-}
-
-
-describe('Custom Context', function() {
-
-  it('should create context from literals', function() {
-
-    const context = EntriesContext.of(15);
-
-    // then
-    expect(context.value.atomicValue).to.exist;
-    expect(context.value.atomicValue).to.equal(15);
-
-  });
-
-});
-
-/**
- * An alternative context that holds additional meta-data
- */
-class EntriesContext extends VariableContext {
-
-  constructor(value = { entries: {} }) {
-    super(value);
-
-    this.value.entries = this.value.entries || {};
-    for (const key in this.value.entries) {
-      const entry = this.value.entries[key];
-
-      const constructor = /** @type { typeof EntriesContext } */ (this.constructor);
-
-      if (!constructor.isAtomic(entry)) {
-        this.value.entries[key] = constructor.of(this.value.entries[key]);
-      }
-    }
-  }
-
-  getKeys() {
-    return Object.keys(this.value.entries);
-  }
-
-  get(key) {
-    return this.value.entries[key];
-  }
-
-  /**
-   * @param {string} key
-   * @param {any} value
-   *
-   * @return {this}
-   */
-  set(key, value) {
-
-    const constructor = /** @type { typeof EntriesContext } */ (this.constructor);
-
-    return /** @type {this} */ (constructor.of(
-      {
-        ...this.value,
-        entries: {
-          ...this.value.entries,
-          [key]: value
-        }
-      }
-    ));
-  }
-
-  static of(...contexts) {
-    const unwrap = (context) => {
-      if (this.isAtomic(context)) {
-        return context instanceof this ?
-          context.value :
-          { atomicValue: context };
-      }
-
-      return { ...context };
-    };
-
-    const merged = contexts.reduce((merged, context) => {
-
-      const {
-        entries = {},
-        ...rest
-      } = unwrap(context);
-
-      return {
-        ...merged,
-        ...rest,
-        entries: {
-          ...merged.entries,
-          ...entries
-        }
-      };
-    }, {});
-
-    return new this(merged);
-  }
-}
-
-
-function toEntriesContextValue(context) {
-
-  return context && Object.keys(context).reduce((result, key) => {
-    const value = context[key];
-
-    result.entries[key] = typeof value === 'object' ? toEntriesContextValue(value)
-      : value;
-
-    return result;
-  }, { entries: {} });
 }
